Move `column-profile` to `web-common/components`
