+++ web-common/src/components/column-profile/ColumnProfileIcon.svelte
