+++ web-common/src/components/column-profile/column-types/NumericProfile.svelte
