+++ web-common/src/components/column-profile/column-types/VarcharProfile.svelte
