+++ web-common/src/components/column-profile/column-types/details/NumericPlot.svelte
