+++ web-common/src/components/column-profile/column-types/details/SummaryNumberPlot.svelte
