+++ web-common/src/components/column-profile/column-types/sparks/NumericSpark.svelte
