+++ web-common/src/components/column-profile/queries.ts
@@ -10,8 +10,8 @@ import {
   V1ProfileColumn,
 } from "@rilldata/web-common/runtime-client";
 import { getPriorityForColumn } from "@rilldata/web-local/lib/http-request-queue/priorities";
+import { convertTimestampPreview } from "@rilldata/web-local/lib/util/convertTimestampPreview";
 import { derived, Readable, writable } from "svelte/store";
-import { convertTimestampPreview } from "../../util/convertTimestampPreview";
 
 export function isFetching(...queries) {
   return queries.some((query) => query?.isFetching);
