+++ web-common/src/features/metrics-views/workspace/MetricsEntityTable.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
+  import ContextButton from "@rilldata/web-common/components/column-profile/ContextButton.svelte";
   import AddIcon from "@rilldata/web-common/components/icons/Add.svelte";
-  import ContextButton from "@rilldata/web-local/lib/components/column-profile/ContextButton.svelte";
   import EditableTable from "@rilldata/web-local/lib/components/table-editable/EditableTable.svelte";
   import { onMount } from "svelte";
