+++ web-common/src/features/models/workspace/inspector/ModelInspectorModelProfile.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+  import ColumnProfile from "@rilldata/web-common/components/column-profile/ColumnProfile.svelte";
   import Shortcut from "@rilldata/web-common/components/tooltip/Shortcut.svelte";
   import TooltipShortcutContainer from "@rilldata/web-common/components/tooltip/TooltipShortcutContainer.svelte";
   import { getFilePathFromNameAndType } from "@rilldata/web-common/features/entity-management/entity-mappers";
@@ -15,7 +16,6 @@
     V1CatalogEntry,
   } from "@rilldata/web-common/runtime-client";
   import { runtimeStore } from "@rilldata/web-local/lib/application-state-stores/application-store";
-  import ColumnProfile from "@rilldata/web-local/lib/components/column-profile/ColumnProfile.svelte";
   import * as classes from "@rilldata/web-local/lib/util/component-classes";
   import { getContext } from "svelte";
   import { derived, Readable, Writable, writable } from "svelte/store";
