+++ web-local/src/lib/components/table-editable/TableRowWithMenu.svelte
@@ -6,7 +6,7 @@
   import { guidGenerator } from "@rilldata/web-common/lib/guid";
   import { createEventDispatcher } from "svelte";
   import { tick } from "svelte/internal";
-  import ContextButton from "../column-profile/ContextButton.svelte";
+  import ContextButton from "../../../../../web-common/src/components/column-profile/ContextButton.svelte";
   import TableHeader from "./TableHeader.svelte";
 
   export let index: number;
